+++ domain/src/main/kotlin/com/gotchai/domain/quiz/entity/AnswerType.kt
@@ -1,6 +1,6 @@
 package com.gotchai.domain.quiz.entity
 
-enum class QuizType {
+enum class AnswerType {
     AI,
     HUMAN,
 }
