+++ storage/rdb/src/main/kotlin/com/gotchai/storage/rdb/quiz/repository/QuizPickJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.gotchai.storage.rdb.quiz.repository
+
+import com.gotchai.storage.rdb.quiz.entity.QuizPickEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface QuizPickJpaRepository : JpaRepository<QuizPickEntity, Long>
